Change C arguments of Fortran INTEGER*8 arguments (back) to long *; add i8_ to all INTEGER*8 functions definitions arguments to make this clear in the future (did not mess with internal variable names)
git-svn-id: file:///home/users/tmap/svn/repos/ferret/trunk@24051 fdbf22ae-c210-0410-be80-ca943da6b8f8

+++ fer/special/FerMem_routines.c
@@ -49,7 +49,7 @@ void *FerMem_Malloc(size_t size, char *filename, int linenumber) {
 
     /* initialize to non-zero junk to catch uninitialized memory usage */
     memset(result, 0x6B, size);
-    sprintf(msg, "%p : 1 : memory malloc allocated for %u bytes : file %s : line %d\n", result, (unsigned int) size, filename, linenumber);
+    sprintf(msg, "%p : 1 : memory malloc allocated for %ld bytes : file %s : line %d\n", result, size, filename, linenumber);
     writedebug(msg);
 #endif
 
@@ -78,7 +78,7 @@ void *FerMem_Realloc(void *ptr, size_t size, char *filename, int linenumber) {
     newptr = realloc(ptr, size);
 
 #ifdef MEMORYDEBUG
-    sprintf(msg, "%p : 3 : memory realloc allocated for %u bytes : file %s : line %d\n", newptr, (unsigned int) size, filename, linenumber);
+    sprintf(msg, "%p : 3 : memory realloc allocated for %ld bytes : file %s : line %d\n", newptr, size, filename, linenumber);
     writedebug(msg);
 #endif
